requirements.txt
asn1crypto==0.24.0
astroid==2.1.0
atomicwrites==1.2.1
attrs==18.2.0
bcrypt==3.1.5
beautifulsoup4==4.6.0
boto3==1.9.82
botocore==1.12.82
certifi==2017.4.17
cffi==1.11.5
chardet==3.0.4
Click==7.0
colorama==0.4.1
colorlog==4.0.2
coverage==4.5.2
cryptography==2.4.2
cssselect==1.0.1
docutils==0.14
feedfinder2==0.0.4
feedparser==5.2.1
Flask==0.12.3
Flask-Bcrypt==0.7.1
Flask-JWT-Extended==3.14.0
Flask-Script==2.0.6
fuzzywuzzy==0.15.0
gunicorn==19.5.0
idna==2.5
isort==4.3.4
itsdangerous==0.24
jieba3k==0.35.1
Jinja2==2.9.6
jmespath==0.9.3
lazy-object-proxy==1.3.1
lxml==4.3.2
MarkupSafe==1.0
mccabe==0.6.1
more-itertools==5.0.0
newspaper3k==0.2.8
nltk==3.4
numpy==1.16.2
pandas==0.24.1
pep8==1.7.1
Pillow==5.4.1
pluggy==0.8.0
psycopg2==2.7.6.1
py==1.7.0
pycparser==2.19
pydbus==0.6.0
PyJWT==1.7.1
pylint==2.2.2
python-Levenshtein
pync==2.0.3
pyOpenSSL==18.0.0
pytest==4.0.2
pytest-cov==2.6.1
python-dateutil==2.7.5
pytz==2018.9
PyYAML==4.2b1
requests==2.20.0
requests-file==1.4.3
s3transfer==0.1.13
scikit-learn==0.20.3
scipy==1.2.1
singledispatch==3.4.0.3
six==1.10.0
summarizer==0.0.7
textblob==0.15.3
tinysegmenter==0.3
tldextract==2.2.1
typed-ast==1.1.1
urllib3==1.23
Werkzeug==0.12.2
wrapt==1.10.11